requirements.txt
imgkit
Pillow
